Separate the text check into an individual check.
